requirements.txt
fastapi
uvicorn
groq
python-dotenv
pydantic
pymongo
motor
python-multipart
python-docx
python-pptx
openpyxl
pandas
aiofiles
python-multipart
supabase
pydantic[email]
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
langchain
langchain-groq
langchain-community
langchain-huggingface
faiss-cpu
pdfplumber
camelot-py[cv]
sentence-transformers
numpy
